get-bad-hosts-yml
Ansible playbook that will retrieve the hostnames.txt and domains.txt files, place them into /etc, change the owner and group to dnsmasq and then restart the dnsmasq daemon.  You need to set the dnsmasq configuration to point to the files in /etc/ or change the ansible playbook to place the files where dnsmasq is configured to look for them.  Playbook should be able to be called from cron. 
conf-file=/etc/domains.txt
addn-hosts=/etc/hostnames.txt